blah

+++ perl/openprint/cron/cleanup.pl
@@ -77,7 +77,7 @@ if ( 1 ) {
 				next;
 			} # end if
 			if ( sql::execute( undef, undef, q{SELECT * FROM tbl_Quote_Details WHERE ProjectIndex=?}, $Project->id() ) ) {
-				$log->error('Quoted!' . $Project->id());
+				#$log->error('Quoted!' . $Project->id());
 				next;
 			} # end if
 			$Project->delete();
@@ -88,8 +88,8 @@ if ( 1 ) {
 	@Projects = openprint::Project::find(
 			'status'=>'Unordered',
 			'order'=>'index desc',
-			'created_on_end' => sprintf('%.4d-%.2d-%.2d', Date::Calc::Add_Delta_Days( Date::Calc::Today(), -365 ) ),
-			'updated_on_end' => sprintf('%.4d-%.2d-%.2d', Date::Calc::Add_Delta_Days( Date::Calc::Today(), -365 ) ),
+			'created_on_end' => sprintf('%.4d-%.2d-%.2d', Date::Calc::Add_Delta_Days( Date::Calc::Today(), -180 ) ),
+			'updated_on_end' => sprintf('%.4d-%.2d-%.2d', Date::Calc::Add_Delta_Days( Date::Calc::Today(), -180 ) ),
 			);
 	if ( @Projects ) {
 		$log->warn("# of Unordered projects to delete: ".@Projects . ' ids ' . $Projects[0]->id() . ' to ' . $Projects[@Projects-1]->id() );
@@ -128,8 +128,9 @@ if ( 1 ) {
 				$log->error('WTF!');
 				next;
 			} # end if
+			next if $Project->docket();
 			if ( sql::execute( undef, undef, q{SELECT * FROM tbl_Quote_Details WHERE ProjectIndex=?}, $Project->id() ) ) {
-				$log->debug('Quoted!' . $Project->id());
+				#$log->debug('Quoted!' . $Project->id());
 				next;
 			} # end if
 			$Project->destroy();
